backend/requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
pillow
python-multipart
spacy>=3.0.0
fpdf2